requirements.txt
aiofiles==0.6.0
asyncpg==0.21.0
boto3==1.15.16
botocore==1.18.16
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
idna==2.10
jmespath==0.10.0
passlib==1.7.4
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-multipart==0.0.5
s3transfer==0.3.3
six==1.15.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
